Split out getGeneratedLocationRanges into a standalone file.
